Move logic to different files
