+++ MercadoPagoSDK/MercadoPagoSDK/MercadoPagoCheckoutScreens.swift
@@ -364,75 +364,3 @@ extension MercadoPagoCheckout {
         onetapFlow.start()
     }
 }
-
-extension OneTapFlow {
-    func showReviewAndConfirmScreenForOneTap() {
-        let reviewVC = PXReviewViewController(viewModel: self.viewModel.reviewConfirmViewModel(), showCustomComponents: false, callbackPaymentData: { [weak self] (paymentData: PaymentData) in
-
-            // One tap
-            if let search = self?.viewModel.search {
-                search.deleteCheckoutDefaultOption()
-            }
-            self?.cancel()
-
-            if !paymentData.hasPaymentMethod() && MercadoPagoCheckoutViewModel.changePaymentMethodCallback != nil {
-                MercadoPagoCheckoutViewModel.changePaymentMethodCallback!()
-            }
-            return
-
-            }, callbackConfirm: {(paymentData: PaymentData) in
-                self.viewModel.updateCheckoutModel(paymentData: paymentData)
-
-                if MercadoPagoCheckoutViewModel.paymentDataConfirmCallback != nil {
-                    MercadoPagoCheckoutViewModel.paymentDataCallback = MercadoPagoCheckoutViewModel.paymentDataConfirmCallback
-                    self.finish()
-                } else {
-                    self.executeNextStep()
-                }
-
-            }, callbackExit: { [weak self] () -> Void in
-                guard let strongSelf = self else {
-                    return
-                }
-                strongSelf.cancel()
-        })
-
-        self.pxNavigationController.pushViewController(viewController: reviewVC, animated: true)
-    }
-
-    func showSecurityCodeScreen() {
-        let securityCodeVc = SecurityCodeViewController(viewModel: self.viewModel.savedCardSecurityCodeViewModel(), collectSecurityCodeCallback: { [weak self] (cardInformation: CardInformationForm, securityCode: String) -> Void in
-            self?.createCardToken(cardInformation: cardInformation as? CardInformation, securityCode: securityCode)
-        })
-        self.pxNavigationController.pushViewController(viewController: securityCodeVc, animated: true)
-    }
-
-    func showPayerCostScreen() {
-        let payerCostStep = AdditionalStepViewController(viewModel: self.viewModel.payerCostViewModel(), callback: { (payerCost) in
-            guard let payerCost = payerCost as? PayerCost else {
-                fatalError("Cannot convert payerCost to type PayerCost")
-            }
-
-            self.viewModel.updateCheckoutModel(payerCost: payerCost)
-            self.executeNextStep()
-        })
-
-        weak var strongPayerCostViewController = payerCostStep
-
-        payerCostStep.viewModel.couponCallback = {[weak self] (discount) in
-            guard let strongSelf = self else {
-                return
-            }
-            strongSelf.viewModel.paymentData.discount = discount
-            strongSelf.getPayerCosts(updateCallback: {
-
-                guard let payerCosts = strongSelf.viewModel.payerCosts, let payerCostViewController = strongPayerCostViewController else {
-                    return
-                }
-                payerCostViewController.updateDataSource(dataSource: payerCosts)
-            })
-
-        }
-        self.pxNavigationController.pushViewController(viewController: payerCostStep, animated: true)
-    }
-}
