+++ MercadoPagoSDK/MercadoPagoSDK/MercadoPagoCheckoutServices.swift
@@ -450,102 +450,3 @@ extension MercadoPagoCheckout {
         })
     }
 }
-
-extension OneTapFlow {
-    func createCardToken(cardInformation: CardInformation? = nil, securityCode: String? = nil) {
-        guard let cardInfo = self.viewModel.paymentOptionSelected as? CardInformation else {
-            return
-        }
-        if self.viewModel.mpESCManager.hasESCEnable() {
-            var savedESCCardToken: SavedESCCardToken
-
-            let esc = self.viewModel.mpESCManager.getESC(cardId: cardInfo.getCardId())
-
-            if !String.isNullOrEmpty(esc) {
-                savedESCCardToken = SavedESCCardToken(cardId: cardInfo.getCardId(), esc: esc)
-            } else {
-                savedESCCardToken = SavedESCCardToken(cardId: cardInfo.getCardId(), securityCode: securityCode)
-            }
-            createSavedESCCardToken(savedESCCardToken: savedESCCardToken)
-
-        } else {
-            createSavedCardToken(cardInformation: cardInfo, securityCode: securityCode!)
-        }
-    }
-
-    func createSavedCardToken(cardInformation: CardInformation, securityCode: String) {
-        self.pxNavigationController.presentLoading()
-
-        let cardInformation = self.viewModel.paymentOptionSelected as! CardInformation
-        let saveCardToken = SavedCardToken(card: cardInformation, securityCode: securityCode, securityCodeRequired: true)
-
-        self.viewModel.mercadoPagoServicesAdapter.createToken(savedCardToken: saveCardToken, callback: { [weak self] (token) in
-
-            if token.lastFourDigits.isEmpty {
-                token.lastFourDigits = cardInformation.getCardLastForDigits()
-            }
-            self?.viewModel.updateCheckoutModel(token: token)
-            self?.executeNextStep()
-
-            }, failure: { [weak self] (error) in
-                let error = MPSDKError.convertFrom(error, requestOrigin: ApiUtil.RequestOrigin.CREATE_TOKEN.rawValue)
-                self?.pxNavigationController.showErrorScreen(error: error, callbackCancel: self?.exitCallback, errorCallback: { [weak self] () in
-                    self?.createSavedCardToken(cardInformation: cardInformation, securityCode: securityCode)
-                })
-        })
-    }
-
-    func createSavedESCCardToken(savedESCCardToken: SavedESCCardToken) {
-        self.pxNavigationController.presentLoading()
-        self.viewModel.mercadoPagoServicesAdapter.createToken(savedESCCardToken: savedESCCardToken, callback: { [weak self] (token) in
-
-            if token.lastFourDigits.isEmpty {
-                let cardInformation = self?.viewModel.paymentOptionSelected as? CardInformation
-                token.lastFourDigits = cardInformation?.getCardLastForDigits() ?? ""
-            }
-            self?.viewModel.updateCheckoutModel(token: token)
-            self?.executeNextStep()
-
-            }, failure: { [weak self] (error) in
-
-                let error = MPSDKError.convertFrom(error, requestOrigin: ApiUtil.RequestOrigin.CREATE_TOKEN.rawValue)
-
-                if let apiException = error.apiException, apiException.containsCause(code: ApiUtil.ErrorCauseCodes.INVALID_ESC.rawValue) ||  apiException.containsCause(code: ApiUtil.ErrorCauseCodes.INVALID_FINGERPRINT.rawValue) {
-
-                    self?.viewModel.mpESCManager.deleteESC(cardId: savedESCCardToken.cardId)
-                }
-
-                self?.pxNavigationController.showErrorScreen(error: error, callbackCancel: self?.exitCallback, errorCallback: { [weak self] () in
-                    self?.createSavedESCCardToken(savedESCCardToken: savedESCCardToken)
-                })
-        })
-    }
-
-    func getPayerCosts(updateCallback: (() -> Void)? = nil) {
-        self.pxNavigationController.presentLoading()
-
-        guard let paymentMethod = self.viewModel.paymentData.getPaymentMethod() else {
-            return
-        }
-
-        self.viewModel.mercadoPagoServicesAdapter.getInstallments(bin: nil, amount: self.viewModel.getAmount(), issuer: self.viewModel.paymentData.getIssuer(), paymentMethodId: paymentMethod.paymentMethodId, callback: { (installments) in
-
-            self.viewModel.payerCosts = installments.first?.payerCosts
-            if let payerCost = installments.first?.payerCosts.first {
-                self.viewModel.updateCheckoutModel(payerCost: payerCost)
-                self.executeNextStep()
-            } else {
-                let error = MPSDKError(message: "Hubo un error".localized, errorDetail: "", retry: true)
-                error.requestOrigin = ApiUtil.RequestOrigin.GET_INSTALLMENTS.rawValue
-                self.pxNavigationController.showErrorScreen(error: error, callbackCancel: self.exitCallback, errorCallback: {
-                    self.getPayerCosts()
-                })
-            }
-        }, failure: {(error) in
-
-            let error = MPSDKError.convertFrom(error, requestOrigin: ApiUtil.RequestOrigin.GET_INSTALLMENTS.rawValue)
-            error.retry = false
-            self.pxNavigationController.showErrorScreen(error: error, callbackCancel: self.exitCallback, errorCallback: nil)
-        })
-    }
-}
