+++ MercadoPagoSDK/MercadoPagoSDK/MercadoPagoCheckoutViewModel.swift
@@ -871,105 +871,3 @@ extension MercadoPagoCheckoutViewModel {
         MercadoPagoCheckoutViewModel.error = nil
     }
 }
-
-class OneTapFlowViewModel: NSObject {
-
-    public enum Steps: String {
-        case ACTION_FINISH
-        case SCREEN_REVIEW_AND_CONFIRM_ONE_TAP
-        case SCREEN_PAYER_COST
-        case SCREEN_SECURITY_CODE
-        case SERVICE_GET_PAYER_COSTS
-    }
-
-    var paymentData: PaymentData
-    let checkoutPreference: CheckoutPreference
-    var paymentOptionSelected: PaymentMethodOption?
-    let search: PaymentMethodSearch
-    var readyToPay: Bool = false
-    var payerCosts: [PayerCost]?
-
-    var mpESCManager: MercadoPagoESC = MercadoPagoESCImplementation()
-    var reviewScreenPreference = ReviewScreenPreference()
-    var mercadoPagoServicesAdapter = MercadoPagoServicesAdapter(servicePreference: MercadoPagoCheckoutViewModel.servicePreference)
-
-    init(paymentData: PaymentData, checkoutPreference: CheckoutPreference, search: PaymentMethodSearch, paymentOptionSelected: PaymentMethodOption) {
-        self.paymentData = paymentData.copy() as! PaymentData
-        self.checkoutPreference = checkoutPreference
-        self.search = search
-        self.paymentOptionSelected = paymentOptionSelected
-    }
-    public func nextStep() -> Steps {
-        if needChosePayerCost() {
-            return .SERVICE_GET_PAYER_COSTS
-        }
-
-        if needPayerCostSelectionScreen() {
-            return .SCREEN_PAYER_COST
-        }
-
-        if needReviewAndConfirmForOneTap() {
-            return .SCREEN_REVIEW_AND_CONFIRM_ONE_TAP
-        }
-
-        if needSecurityCode() {
-            return .SCREEN_SECURITY_CODE
-        }
-        return .ACTION_FINISH
-    }
-}
-
-extension OneTapFlowViewModel {
-
-    public func savedCardSecurityCodeViewModel() -> SecurityCodeViewModel {
-        guard let cardInformation = self.paymentOptionSelected as? CardInformation else {
-            fatalError("Cannot conver payment option selected to CardInformation")
-        }
-
-        let reason = SecurityCodeViewModel.Reason.SAVED_CARD
-        return SecurityCodeViewModel(paymentMethod: self.paymentData.paymentMethod!, cardInfo: cardInformation, reason: reason)
-    }
-
-    func reviewConfirmViewModel() -> PXReviewViewModel {
-        return PXReviewViewModel(checkoutPreference: self.checkoutPreference, paymentData: self.paymentData, paymentOptionSelected: self.paymentOptionSelected!, discount: paymentData.discount, reviewScreenPreference: reviewScreenPreference)
-    }
-
-    public func payerCostViewModel() -> AdditionalStepViewModel {
-        var paymentMethod: PaymentMethod = PaymentMethod()
-        if let pm = self.paymentData.getPaymentMethod() {
-            paymentMethod = pm
-        }
-
-        guard let cardInformation = paymentOptionSelected as? CardInformationForm else {
-            fatalError("Cannot conver payment option selected to CardInformation")
-        }
-
-        return PayerCostAdditionalStepViewModel(amount: self.getAmount(), token: cardInformation, paymentMethod: paymentMethod, dataSource: payerCosts!, discount: self.paymentData.discount, email: self.checkoutPreference.payer.email, mercadoPagoServicesAdapter: mercadoPagoServicesAdapter)
-    }
-}
-
-extension OneTapFlowViewModel {
-    func updateCheckoutModel(paymentData: PaymentData) {
-        self.paymentData = paymentData
-        self.readyToPay = true
-    }
-
-    public func updateCheckoutModel(token: Token) {
-        self.paymentData.updatePaymentDataWith(token: token)
-    }
-
-    public func updateCheckoutModel(payerCost: PayerCost) {
-        self.paymentData.updatePaymentDataWith(payerCost: payerCost)
-        self.paymentData.cleanToken()
-    }
-
-    internal func getAmount() -> Double {
-        if let payerCost = paymentData.getPayerCost() {
-            return payerCost.totalAmount
-        } else if MercadoPagoCheckoutViewModel.flowPreference.isDiscountEnable(), let discount = paymentData.discount {
-            return discount.newAmount()
-        } else {
-            return self.checkoutPreference.getAmount()
-        }
-    }
-}
