+++ MercadoPagoSDK/MercadoPagoSDK/NextStepHelper.swift
@@ -395,77 +395,3 @@ extension MercadoPagoCheckoutViewModel {
          return MercadoPagoCheckoutViewModel.flowPreference.isDiscountEnable()
     }
 }
-
-extension OneTapFlowViewModel {
-    func needReviewAndConfirmForOneTap() -> Bool {
-        guard self.paymentOptionSelected != nil else {
-            return false
-        }
-
-        if readyToPay {
-            return false
-        }
-
-        if paymentData.isComplete(shouldCheckForToken: false) {
-            return true
-        }
-
-        if paymentData.isComplete() {
-            return MercadoPagoCheckoutViewModel.flowPreference.isReviewAndConfirmScreenEnable()
-        }
-        return false
-    }
-
-    func needSecurityCode() -> Bool {
-        guard let pmSelected = self.paymentOptionSelected else {
-            return false
-        }
-
-        guard let pm = self.paymentData.getPaymentMethod() else {
-            return false
-        }
-
-        if !readyToPay {
-            return false
-        }
-
-        let hasInstallmentsIfNeeded = paymentData.hasPayerCost() || !pm.isCreditCard
-        let isCustomerCard = pmSelected.isCustomerPaymentMethod() && pmSelected.getId() != PaymentTypeId.ACCOUNT_MONEY.rawValue
-
-        if  isCustomerCard && !paymentData.hasToken() && hasInstallmentsIfNeeded && !hasSavedESC() {
-            return true
-        }
-        return false
-    }
-
-    func hasSavedESC() -> Bool {
-        guard let pmSelected = self.paymentOptionSelected else {
-            return false
-        }
-
-        if let card = pmSelected as? CardInformation {
-            return mpESCManager.getESC(cardId: card.getCardId()) == nil ? false : true
-        }
-        return false
-    }
-
-    func needChosePayerCost() -> Bool {
-        guard let pm = self.paymentData.getPaymentMethod() else {
-            return false
-        }
-        if pm.isCreditCard && !paymentData.hasPayerCost() && payerCosts == nil {
-            return true
-        }
-        return false
-    }
-
-    func needPayerCostSelectionScreen() -> Bool {
-        guard let pm = self.paymentData.getPaymentMethod() else {
-            return false
-        }
-        if pm.isCreditCard && !paymentData.hasPayerCost() && payerCosts != nil {
-            return true
-        }
-        return false
-    }
-}
